requirements.txt
Django==1.11.7
djangorestframework==3.7.3
Faker==0.8.7